+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-10T08:46:05.452Z",
+    "2026-06-10T10:47:46.936Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {